requirements.txt
Flask==1.1.4
flask-restplus==0.13.0
Werkzeug==0.16.1
keras==2.6.0
Keras-Preprocessing==1.1.2
keras-vis==0.4.1
opencv-python==4.4.0.46
pandas==1.1.5
pandas-datareader==0.9.0
pathlib==1.0.1
Pillow==8.4.0
scikit-image==0.18.3
tensorflow-cpu
urllib3==1.24.3